webhook_payment/requirements.txt
fastapi
uvicorn[standard]
databases
asyncpg
sqlalchemy
pydantic
python-dotenv
